Completable future task

+++ src/main/java/com/example/rest/Employee.java
@@ -3,12 +3,12 @@ package com.example.rest;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Data
 @Builder
 public class Employee {
     private String name;
     private Long id;
-    private BigInteger salary;
+    private BigDecimal salary;
 }
